Unify GraphHistory
