+++ optuna_dashboard/ts/components/GraphHistory.tsx
@@ -17,18 +17,24 @@ import {
 } from "@mui/material"
 import { plotlyDarkTemplate } from "./PlotlyDarkMode"
 import {
-  useFilteredTrials,
+  useFilteredTrialsFromStudies,
   Target,
   useObjectiveAndUserAttrTargets,
 } from "../trialFilter"
 
 const plotDomId = "graph-history"
 
+interface HistoryPlotInfo {
+  study_name: string
+  trials: Trial[]
+  directions: StudyDirection[]
+}
+
 export const GraphHistory: FC<{
-  study: StudyDetail | null
+  studies: StudyDetail[] | null
   betaLogScale?: boolean
   betaIncludePruned?: boolean
-}> = ({ study, betaLogScale, betaIncludePruned }) => {
+}> = ({ studies, betaLogScale, betaIncludePruned }) => {
   const theme = useTheme()
   const [xAxis, setXAxis] = useState<
     "number" | "datetime_start" | "datetime_complete"
@@ -37,29 +43,30 @@ export const GraphHistory: FC<{
   const [filterCompleteTrial, setFilterCompleteTrial] = useState<boolean>(false)
   const [filterPrunedTrial, setFilterPrunedTrial] = useState<boolean>(false)
 
-  const [targets, selected, setTarget] = useObjectiveAndUserAttrTargets(study)
-  const trials = useFilteredTrials(
-    study,
-    [selected],
-    filterCompleteTrial,
-    betaIncludePruned === undefined ? filterPrunedTrial : !betaIncludePruned
+  const [targets, selected, setTarget] = useObjectiveAndUserAttrTargets(studies ? studies[0] : null)
+
+  const trials = useFilteredTrialsFromStudies(studies, [selected], filterCompleteTrial, betaIncludePruned === undefined ? filterPrunedTrial : !betaIncludePruned
   )
+  const historyPlotInfos = studies.map((study, index) => {
+    const h: HistoryPlotInfo = {
+      study_name: study?.name,
+      trials: trials[index],
+      directions: study?.directions,
+    }
+    return h
+  })
 
   useEffect(() => {
-    if (study !== null) {
-      plotHistory(
-        trials,
-        study.directions,
-        selected,
-        xAxis,
-        betaLogScale === undefined ? logScale : betaLogScale,
-        theme.palette.mode,
-        study?.objective_names
-      )
-    }
+    plotHistory(
+      historyPlotInfos,
+      selected,
+      xAxis,
+      betaLogScale === undefined ? logScale : betaLogScale,
+      theme.palette.mode
+      study?.objective_names
+    )
   }, [
-    trials,
-    study?.directions,
+    studies,
     selected,
     logScale,
     betaLogScale,
@@ -109,7 +116,7 @@ export const GraphHistory: FC<{
         >
           History
         </Typography>
-        {targets.length >= 2 ? (
+        {studies[0] !== null && targets.length >= 2 ? (
           <FormControl
             component="fieldset"
             sx={{ marginBottom: theme.spacing(2) }}
@@ -121,7 +128,7 @@ export const GraphHistory: FC<{
             >
               {targets.map((t, i) => (
                 <MenuItem value={t.identifier()} key={i}>
-                  {t.toLabel(study?.objective_names)}
+                  {t.toLabel(studies[0].objective_names)}
                 </MenuItem>
               ))}
             </Select>
@@ -159,7 +166,7 @@ export const GraphHistory: FC<{
               control={
                 <Checkbox
                   checked={!filterPrunedTrial}
-                  disabled={!study?.has_intermediate_values}
+                  disabled={!studies[0]?.has_intermediate_values}
                   onChange={handleFilterPrunedChange}
                 />
               }
@@ -204,8 +211,7 @@ export const GraphHistory: FC<{
 }
 
 const plotHistory = (
-  trials: Trial[],
-  directions: StudyDirection[],
+  historyPlotInfos: HistoryPlotInfo[],
   target: Target,
   xAxis: "number" | "datetime_start" | "datetime_complete",
   logScale: boolean,
@@ -213,6 +219,12 @@ const plotHistory = (
   objectiveNames?: string[]
 ) => {
   if (document.getElementById(plotDomId) === null) {
+    return
+  }
+  if (historyPlotInfos.length === 0) {
+    plotly.react(domId, [], {
+      template: mode === "dark" ? plotlyDarkTemplate : {},
+    })
     return
   }
 
@@ -234,10 +246,6 @@ const plotHistory = (
     showlegend: true,
     template: mode === "dark" ? plotlyDarkTemplate : {},
   }
-  if (trials.length === 0) {
-    plotly.react(plotDomId, [], layout)
-    return
-  }
 
   const getAxisX = (trial: Trial): number | Date => {
     return xAxis === "number"
@@ -247,62 +255,73 @@ const plotHistory = (
       : trial.datetime_complete!
   }
 
-  const plotData: Partial<plotly.PlotData>[] = [
-    {
-      x: trials.map(getAxisX),
-      y: trials.map((t: Trial): number => target.getTargetValue(t) as number),
-      name: target.toLabel(objectiveNames),
+  const plotData: Partial<plotly.PlotData>[] = []
+  historyPlotInfos.forEach((h) => {
+    const x = h.trials.map(getAxisX)
+    const y = h.trials.map(
+      (t: Trial): number => target.getTargetValue(t) as number
+    )
+    plotData.push({
+      x: x,
+      y: y,
+      name: `${target.toLabel(objectiveNames)} of ${h.study_name}`,
       mode: "markers",
       type: "scatter",
-    },
-  ]
-
-  const objectiveId = target.getObjectiveId()
-  if (objectiveId !== null) {
-    const xForLinePlot: (number | Date)[] = []
-    const yForLinePlot: number[] = []
-    let currentBest: number | null = null
-    for (let i = 0; i < trials.length; i++) {
-      const t = trials[i]
-      if (currentBest === null) {
-        currentBest = t.values![objectiveId] as number
-        xForLinePlot.push(getAxisX(t))
-        yForLinePlot.push(t.values![objectiveId] as number)
-      } else if (
-        directions[objectiveId] === "maximize" &&
-        t.values![objectiveId] > currentBest
-      ) {
-        const p = trials[i - 1]
-        if (!xForLinePlot.includes(getAxisX(p))) {
-          xForLinePlot.push(getAxisX(p))
-          yForLinePlot.push(currentBest)
-        }
-        currentBest = t.values![objectiveId] as number
-        xForLinePlot.push(getAxisX(t))
-        yForLinePlot.push(t.values![objectiveId] as number)
-      } else if (
-        directions[objectiveId] === "minimize" &&
-        t.values![objectiveId] < currentBest
-      ) {
-        const p = trials[i - 1]
-        if (!xForLinePlot.includes(getAxisX(p))) {
-          xForLinePlot.push(getAxisX(p))
-          yForLinePlot.push(currentBest)
-        }
-        currentBest = t.values![objectiveId] as number
-        xForLinePlot.push(getAxisX(t))
-        yForLinePlot.push(t.values![objectiveId] as number)
-      }
-    }
-    xForLinePlot.push(getAxisX(trials[trials.length - 1]))
-    yForLinePlot.push(yForLinePlot[yForLinePlot.length - 1])
-    plotData.push({
-      x: xForLinePlot,
-      y: yForLinePlot,
-      name: "Best Value",
-      mode: "lines",
-      type: "scatter",
     })
-  }
+
+    const objectiveId = target.getObjectiveId()
+    if (objectiveId !== null) {
+      const xForLinePlot: (number | Date)[] = []
+      const yForLinePlot: number[] = []
+      let currentBest: number | null = null
+      for (let i = 0; i < h.trials.length; i++) {
+        const t = h.trials[i]
+        const value = target.getTargetValue(t) as number
+        if (value === null) {
+          continue
+        } else if (currentBest === null) {
+          currentBest = value
+          xForLinePlot.push(getAxisX(t))
+          yForLinePlot.push(value)
+        } else if (
+          h.directions[objectiveId] === "maximize" &&
+          value > currentBest
+        ) {
+          const p = h.trials[i - 1]
+          if (!xForLinePlot.includes(getAxisX(p))) {
+            xForLinePlot.push(getAxisX(p))
+            yForLinePlot.push(currentBest)
+          }
+          currentBest = value
+          xForLinePlot.push(getAxisX(t))
+          yForLinePlot.push(value)
+        } else if (
+          h.directions[objectiveId] === "minimize" &&
+          value < currentBest
+        ) {
+          const p = h.trials[i - 1]
+          if (!xForLinePlot.includes(getAxisX(p))) {
+            xForLinePlot.push(getAxisX(p))
+            yForLinePlot.push(currentBest)
+          }
+          currentBest = value
+          xForLinePlot.push(getAxisX(t))
+          yForLinePlot.push(value)
+        }
+      }
+      if (h.trials.length !== 0) {
+        xForLinePlot.push(getAxisX(h.trials[h.trials.length - 1]))
+        yForLinePlot.push(yForLinePlot[yForLinePlot.length - 1])
+      }
+      plotData.push({
+        x: xForLinePlot,
+        y: yForLinePlot,
+        name: `Best Value of ${h.study_name}`,
+        mode: "lines",
+        type: "scatter",
+      })
+    }
+  })
+
   plotly.react(plotDomId, plotData, layout)
 }
