+++ optuna_dashboard/ts/components/StudyHistory.tsx
@@ -95,7 +95,7 @@ export const StudyHistory: FC<{ studyId: number }> = ({ studyId }) => {
       >
         <CardContent>
           <GraphHistory
-            study={studyDetail}
+            studies={[studyDetail, ]}
             betaIncludePruned={includePruned}
             betaLogScale={logScale}
           />
