+++ optuna_dashboard/ts/trialFilter.ts
@@ -116,7 +116,7 @@ export const useFilteredTrials = (
     })
   }, [study?.trials, targets, filterComplete, filterPruned])
 
-export const useFilteredTrials1 = (
+export const useFilteredTrialsFromStudies = (
   studies: StudyDetail[],
   targets: Target[],
   filterComplete: boolean,
@@ -141,29 +141,6 @@ export const useFilteredTrials1 = (
       })
     })
   }, [studies, targets, filterComplete, filterPruned])
-
-export const getFilteredTrials = (
-  study: StudyDetail | null,
-  targets: Target[],
-  filterComplete: boolean,
-  filterPruned: boolean
-): Trial[] => {
-  if (study === null) {
-    return []
-  }
-  return study.trials.filter((t) => {
-    if (t.state !== "Complete" && t.state !== "Pruned") {
-      return false
-    }
-    if (t.state === "Complete" && filterComplete) {
-      return false
-    }
-    if (t.state === "Pruned" && filterPruned) {
-      return false
-    }
-    return targets.every((target) => !target.getTargetValue(t) !== null)
-  })
-}
 
 export const useObjectiveTargets = (
   study: StudyDetail | null
